Webapp/requirements.txt
Flask==1.1.2
gunicorn==19.9.0
Pillow==6.0.0
opencv-python-headless==4.2.0.34
numpy==1.19.5
scikit-image==0.18.0rc1
scikit-learn==0.22.2.post1
scipy==1.2.1
tensorflow==1.13.1
